refactor, removing things from manager
