+++ conans/client/cmd/test.py
@@ -4,6 +4,7 @@ import tempfile
 
 from conans.util.files import rmdir
 from conans.util.env_reader import get_env
+from conans.client.cmd.build import build
 
 
 class PackageTester(object):
@@ -21,7 +22,8 @@ class PackageTester(object):
         and builds the test_package/conanfile.py running the test() method.
         """
         base_folder = os.path.dirname(conanfile_abs_path)
-        test_build_folder, delete_after_build = self._build_folder(test_build_folder, profile, base_folder)
+        test_build_folder, delete_after_build = self._build_folder(test_build_folder, profile,
+                                                                   base_folder)
         rmdir(test_build_folder)
         if build_modes is None:
             build_modes = ["never"]
@@ -37,11 +39,14 @@ class PackageTester(object):
                                   manifest_verify=manifest_verify,
                                   manifest_interactive=manifest_interactive,
                                   keep_build=keep_build)
-            self._manager.build(conanfile_abs_path, base_folder, test_build_folder, package_folder=None,
-                                install_folder=test_build_folder, test=str(reference))
+            # FIXME: This is ugly access to graph_manager and plugin_manager. Will be cleaned in 2.0
+            build(self._manager._graph_manager, self._manager._plugin_manager, conanfile_abs_path,
+                  self._user_io.out,
+                  base_folder, test_build_folder, package_folder=None,
+                  install_folder=test_build_folder, test=str(reference))
         finally:
             if delete_after_build:
-                os.chdir(base_folder)    # Required for windows where deleting the cwd is not possible.
+                os.chdir(base_folder)  # Required for windows where deleting the cwd is not possible.
                 rmdir(test_build_folder)
 
     @staticmethod
