+++ conans/client/manager.py
@@ -1,23 +1,19 @@
 import os
 
-from conans.client import packager
 from conans.client.client_cache import ClientCache
-from conans.client.cmd.export import _execute_export
 from conans.client.generators import write_generators
 from conans.client.importer import run_imports, run_deploy
 from conans.client.installer import ConanInstaller, call_system_requirements
 from conans.client.manifest_manager import ManifestManager
 from conans.client.output import ScopedOutput, Color
-from conans.client.source import config_source_local, complete_recipe_sources
+from conans.client.source import complete_recipe_sources
 from conans.client.tools import cross_building, get_cross_building_settings
 from conans.client.userio import UserIO
-from conans.errors import NotFoundException, ConanException, conanfile_exception_formatter
-from conans.model.conan_file import get_env_context_manager
-from conans.model.ref import ConanFileReference, PackageReference
-from conans.paths import CONANFILE, CONANINFO, CONANFILE_TXT, BUILD_INFO
-from conans.util.files import save, rmdir, normalize, mkdir
-from conans.util.log import logger
-from conans.client.graph.graph_manager import load_deps_info
+from conans.errors import ConanException
+from conans.model.ref import ConanFileReference
+from conans.paths import CONANINFO
+from conans.util.files import save, normalize
+
 from conans.client.graph.printer import print_graph
 
 
@@ -37,46 +33,6 @@ class ConanManager(object):
         self._registry = registry
         self._graph_manager = graph_manager
         self._plugin_manager = plugin_manager
-
-    def export_pkg(self, reference, source_folder, build_folder, package_folder, install_folder,
-                   profile, force):
-
-        conan_file_path = self._client_cache.conanfile(reference)
-        if not os.path.exists(conan_file_path):
-            raise ConanException("Package recipe '%s' does not exist" % str(reference))
-
-        deps_graph = self._graph_manager.load_simple_graph(reference, profile, self._recorder)
-
-        # this is a bit tricky, but works. The root (virtual), has only 1 neighbor,
-        # which is the exported pkg
-        nodes = deps_graph.root.neighbors()
-        conanfile = nodes[0].conanfile
-        if install_folder and existing_info_files(install_folder):
-            load_deps_info(install_folder, conanfile, required=True)
-        pkg_id = conanfile.info.package_id()
-        self._user_io.out.info("Packaging to %s" % pkg_id)
-        pkg_reference = PackageReference(reference, pkg_id)
-        dest_package_folder = self._client_cache.package(pkg_reference,
-                                                         short_paths=conanfile.short_paths)
-
-        if os.path.exists(dest_package_folder):
-            if force:
-                rmdir(dest_package_folder)
-            else:
-                raise ConanException("Package already exists. Please use --force, -f to "
-                                     "overwrite it")
-
-        recipe_hash = self._client_cache.load_manifest(reference).summary_hash
-        conanfile.info.recipe_hash = recipe_hash
-        conanfile.develop = True
-        package_output = ScopedOutput(str(reference), self._user_io.out)
-        if package_folder:
-            packager.export_pkg(conanfile, pkg_id, package_folder, dest_package_folder,
-                                package_output, self._plugin_manager, conan_file_path, reference)
-        else:
-            packager.create_package(conanfile, pkg_id, source_folder, build_folder,
-                                    dest_package_folder, install_folder, package_output,
-                                    self._plugin_manager, conan_file_path, reference, local=True)
 
     def install_workspace(self, profile, workspace, remote_name, build_modes, update):
         references = [ConanFileReference(v, "root", "project", "develop") for v in workspace.root]
@@ -184,110 +140,3 @@ class ConanManager(object):
                 deploy_conanfile = neighbours[0].conanfile
                 if hasattr(deploy_conanfile, "deploy") and callable(deploy_conanfile.deploy):
                     run_deploy(deploy_conanfile, install_folder, output)
-
-    def source(self, conanfile_path, source_folder, info_folder):
-        """
-        :param conanfile_path: Absolute path to a conanfile
-        :param source_folder: Absolute path where to put the files
-        :param info_folder: Absolute path where to read the info files
-        :param package_folder: Absolute path to the package_folder, only to have the var present
-        :return:
-        """
-        output = ScopedOutput("PROJECT", self._user_io.out)
-        # only infos if exist
-        conanfile = self._graph_manager.load_consumer_conanfile(conanfile_path, info_folder, output)
-        conanfile_folder = os.path.dirname(conanfile_path)
-        if conanfile_folder != source_folder:
-            output.info("Executing exports to: %s" % source_folder)
-            _execute_export(conanfile_path, conanfile, source_folder, source_folder, output)
-        config_source_local(source_folder, conanfile, conanfile_folder, output, conanfile_path,
-                            self._plugin_manager)
-
-    def imports(self, conan_file_path, dest_folder, info_folder):
-        """
-        :param conan_file_path: Abs path to a conanfile
-        :param dest_folder:  Folder where to put the files
-        :param info_folder: Folder containing the conaninfo/conanbuildinfo.txt files
-        :return:
-        """
-
-        output = ScopedOutput("PROJECT", self._user_io.out)
-        conanfile = self._graph_manager.load_consumer_conanfile(conan_file_path, info_folder, output,
-                                                                deps_info_required=True)
-        run_imports(conanfile, dest_folder, output)
-
-    def local_package(self, package_folder, conanfile_path, build_folder, source_folder,
-                      install_folder):
-        if package_folder == build_folder:
-            raise ConanException("Cannot 'conan package' to the build folder. "
-                                 "--build-folder and package folder can't be the same")
-        output = ScopedOutput("PROJECT", self._user_io.out)
-        conanfile = self._graph_manager.load_consumer_conanfile(conanfile_path, install_folder,
-                                                                output, deps_info_required=True)
-        packager.create_package(conanfile, None, source_folder, build_folder, package_folder,
-                                install_folder, output, self._plugin_manager, conanfile_path, None,
-                                local=True, copy_info=True)
-
-    def build(self, conanfile_path, source_folder, build_folder, package_folder, install_folder,
-              test=False, should_configure=True, should_build=True, should_install=True,
-              should_test=True):
-        """ Call to build() method saved on the conanfile.py
-        param conanfile_path: path to a conanfile.py
-        """
-        logger.debug("Building in %s" % build_folder)
-        logger.debug("Conanfile in %s" % conanfile_path)
-
-        try:
-            # Append env_vars to execution environment and clear when block code ends
-            output = ScopedOutput(("%s (test package)" % test) if test else "Project",
-                                  self._user_io.out)
-            conan_file = self._graph_manager.load_consumer_conanfile(conanfile_path, install_folder,
-                                                                     output, deps_info_required=True)
-        except NotFoundException:
-            # TODO: Auto generate conanfile from requirements file
-            raise ConanException("'%s' file is needed for build.\n"
-                                 "Create a '%s' and move manually the "
-                                 "requirements and generators from '%s' file"
-                                 % (CONANFILE, CONANFILE, CONANFILE_TXT))
-
-        if test:
-            try:
-                conan_file.requires.add(test)
-            except ConanException:
-                pass
-
-        conan_file.should_configure = should_configure
-        conan_file.should_build = should_build
-        conan_file.should_install = should_install
-        conan_file.should_test = should_test
-
-        try:
-            mkdir(build_folder)
-            os.chdir(build_folder)
-            conan_file.build_folder = build_folder
-            conan_file.source_folder = source_folder
-            conan_file.package_folder = package_folder
-            conan_file.install_folder = install_folder
-            self._plugin_manager.execute("pre_build", conanfile=conan_file,
-                                         conanfile_path=conanfile_path)
-            with get_env_context_manager(conan_file):
-                output.highlight("Running build()")
-                with conanfile_exception_formatter(str(conan_file), "build"):
-                    conan_file.build()
-                self._plugin_manager.execute("post_build", conanfile=conan_file,
-                                             conanfile_path=conanfile_path)
-                if test:
-                    output.highlight("Running test()")
-                    with conanfile_exception_formatter(str(conan_file), "test"):
-                        conan_file.test()
-        except ConanException:
-            raise  # Raise but not let to reach the Exception except (not print traceback)
-        except Exception:
-            import traceback
-            trace = traceback.format_exc().split('\n')
-            raise ConanException("Unable to build it successfully\n%s" % '\n'.join(trace[3:]))
-
-
-def existing_info_files(folder):
-    return os.path.exists(os.path.join(folder, CONANINFO)) and  \
-           os.path.exists(os.path.join(folder, BUILD_INFO))
