+++ conans/test/command/upload_complete_test.py
@@ -1,19 +1,21 @@
+import os
 import json
 import unittest
+import platform
+import stat
+
+from requests.packages.urllib3.exceptions import ConnectionError
+
 from conans.test.utils.tools import TestClient, TestServer, TestRequester
 from conans.test.utils.test_files import hello_source_files, temp_folder,\
     hello_conan_files
-from conans.client.manager import CONANFILE
-import os
-from conans.paths import CONAN_MANIFEST, EXPORT_TGZ_NAME, CONANINFO
-import platform
-import stat
+from conans.paths import CONAN_MANIFEST, EXPORT_TGZ_NAME, CONANINFO, CONANFILE
 from conans.util.files import load, mkdir, save
 from conans.model.ref import ConanFileReference, PackageReference
 from conans.model.manifest import FileTreeManifest
 from conans.test.utils.test_files import uncompress_packaged_files
 from conans.tools import untargz
-from requests.packages.urllib3.exceptions import ConnectionError
+
 from conans.test.utils.cpp_test_files import cpp_hello_conan_files
 
 myconan1 = """
@@ -169,7 +171,7 @@ class UploadTest(unittest.TestCase):
         client.run("export . frodo/stable")
         client.run("upload Hello* --confirm --retry 10 --retry-wait=0", ignore_error=True)
         self.assertIn("Waiting 0 seconds to retry...", client.user_io.out)
-        self.assertIn("ERROR: Execute upload again to retry upload the failed files", client.user_io.out)
+        self.assertIn("ERROR: Execute upload again to retry upload the failed files", client.out)
 
         # For each file will fail the first time and will success in the second one
         client = self._get_client(FailPairFilesUploader)
